Completed authorization

+++ nestjs-task-management/src/tasks/tasks.service.ts
@@ -21,8 +21,8 @@ export class TasksService {
         return await this.taskRepository.getTask(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
         if (!found){
             throw new NotFoundException(`There is no task with id: ${id}`);
         }
@@ -36,15 +36,21 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto, user)
     }
 
-    async deleteTaskById(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTaskById(
+        id: number,
+        user: User
+        ): Promise<void> {
+        const result = await this.taskRepository.delete({ id , userId: user.id });
         if (result.affected === 0 ){
             throw new NotFoundException(`Task with id: ${id}, not found`)
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
